feat(job): 增加一致性hash

+++ rabbit-job-admin/src/main/java/com/snail/job/admin/model/JobLog.java
@@ -28,7 +28,7 @@ public class JobLog {
     /**
      * 任务，主键ID
      */
-    private Integer jobId;
+    private Long jobId;
 
     /**
      * 任务组名
