+++ rabbit-job-common/src/main/java/com/snail/job/common/enums/AlarmStatus.java
@@ -10,24 +10,24 @@ public enum AlarmStatus {
     /**
      * 无需告警
      */
-    NO_ALARM("no_alarm", (byte) 0, "无需告警"),
+    NO_ALARM("no_alarm", 0, "无需告警"),
 
     /**
      * 待告警
      */
-    WAIT_ALARM("wait_alarm", (byte) 1, "待告警"),
+    WAIT_ALARM("wait_alarm", 1, "待告警"),
 
     /**
      * 已告警
      */
-    ALARM_FINISH("alarm_finish", (byte) 2, "已告警"),
+    ALARM_FINISH("alarm_finish", 2, "已告警"),
     ;
 
     private final String name;
-    private final Byte value;
+    private final Integer value;
     private final String desc;
 
-    AlarmStatus(String name, Byte value, String desc) {
+    AlarmStatus(String name, Integer value, String desc) {
         this.name = name;
         this.value = value;
         this.desc = desc;
@@ -37,7 +37,7 @@ public enum AlarmStatus {
         return name;
     }
 
-    public Byte getValue() {
+    public Integer getValue() {
         return value;
     }
